fix: rename to dataplatform
