+++ docs/api/dataplatform/dataplatform_cluster.md
@@ -1,4 +1,4 @@
-# data_platform_cluster
+# dataplatform_cluster
 
 This is a simple module that supports creating or removing Data Platform Clusters. This module has a dependency on ionoscloud &gt;= 6.0.2
 
@@ -8,23 +8,23 @@ This is a simple module that supports creating or removing Data Platform Cluster
 ```yaml
 
   - name: Create Data Platform cluster
-    data_platform_cluster:
+    dataplatform_cluster:
       name: "{{ cluster_name }}"
   
 
   - name: Update Data Platform cluster
-    data_platform_cluster:
-      data_platform_cluster_id: "89a5aeb0-d6c1-4cef-8f6b-2b9866d85850"
+    dataplatform_cluster:
+      dataplatform_cluster_id: "89a5aeb0-d6c1-4cef-8f6b-2b9866d85850"
       maintenance_window:
         day_of_the_week: 'Tuesday'
         time: '13:03:00'
-      data_platform_version: 1.17.8
+      dataplatform_version: 1.17.8
       state: update
   
 
   - name: Delete Data Platform cluster
-    data_platform_cluster:
-      data_platform_cluster_id: "a9b56a4b-8033-4f1a-a59d-cfea86cfe40b"
+    dataplatform_cluster:
+      dataplatform_cluster_id: "a9b56a4b-8033-4f1a-a59d-cfea86cfe40b"
       state: absent
   
 ```
@@ -36,7 +36,7 @@ This is a simple module that supports creating or removing Data Platform Cluster
 ```yaml
   
   - name: Create Data Platform cluster
-    data_platform_cluster:
+    dataplatform_cluster:
       name: "{{ cluster_name }}"
   
 ```
@@ -46,7 +46,7 @@ This is a simple module that supports creating or removing Data Platform Cluster
   | Name | Required | Type | Default | Description |
   | :--- | :---: | :--- | :--- | :--- |
   | cluster_name | True | str |  | The name of your cluster. Must be 63 characters or less and must be empty or begin and end with an alphanumeric character ([a-z0-9A-Z]) with dashes (-), underscores (_), dots (.), and alphanumerics between. |
-  | data_platform_version | False | str |  | The version of the DataPlatform. |
+  | dataplatform_version | False | str |  | The version of the DataPlatform. |
   | datacenter_id | True | str |  | The UUID of the virtual data center (VDC) the cluster is provisioned. |
   | maintenance_window | False | dict |  | Starting time of a weekly 4 hour-long window, during which maintenance might occur in hh:mm:ss format |
   | api_url | False | str |  | The Ionos API base URL. |
@@ -64,8 +64,8 @@ This is a simple module that supports creating or removing Data Platform Cluster
 ```yaml
   
   - name: Delete Data Platform cluster
-    data_platform_cluster:
-      data_platform_cluster_id: "a9b56a4b-8033-4f1a-a59d-cfea86cfe40b"
+    dataplatform_cluster:
+      dataplatform_cluster_id: "a9b56a4b-8033-4f1a-a59d-cfea86cfe40b"
       state: absent
   
 ```
@@ -74,7 +74,7 @@ This is a simple module that supports creating or removing Data Platform Cluster
 
   | Name | Required | Type | Default | Description |
   | :--- | :---: | :--- | :--- | :--- |
-  | data_platform_cluster_id | True | str |  | The ID of the Data Platform cluster. |
+  | dataplatform_cluster_id | True | str |  | The ID of the Data Platform cluster. |
   | api_url | False | str |  | The Ionos API base URL. |
   | username | False | str |  | The Ionos username. Overrides the IONOS_USERNAME environment variable. |
   | password | False | str |  | The Ionos password. Overrides the IONOS_PASSWORD environment variable. |
@@ -90,12 +90,12 @@ This is a simple module that supports creating or removing Data Platform Cluster
 ```yaml
   
   - name: Update Data Platform cluster
-    data_platform_cluster:
-      data_platform_cluster_id: "89a5aeb0-d6c1-4cef-8f6b-2b9866d85850"
+    dataplatform_cluster:
+      dataplatform_cluster_id: "89a5aeb0-d6c1-4cef-8f6b-2b9866d85850"
       maintenance_window:
         day_of_the_week: 'Tuesday'
         time: '13:03:00'
-      data_platform_version: 1.17.8
+      dataplatform_version: 1.17.8
       state: update
   
 ```
@@ -105,8 +105,8 @@ This is a simple module that supports creating or removing Data Platform Cluster
   | Name | Required | Type | Default | Description |
   | :--- | :---: | :--- | :--- | :--- |
   | cluster_name | True | str |  | The name of your cluster. Must be 63 characters or less and must be empty or begin and end with an alphanumeric character ([a-z0-9A-Z]) with dashes (-), underscores (_), dots (.), and alphanumerics between. |
-  | data_platform_cluster_id | True | str |  | The ID of the Data Platform cluster. |
-  | data_platform_version | True | str |  | The version of the DataPlatform. |
+  | dataplatform_cluster_id | True | str |  | The ID of the Data Platform cluster. |
+  | dataplatform_version | True | str |  | The version of the DataPlatform. |
   | datacenter_id | False | str |  | The UUID of the virtual data center (VDC) the cluster is provisioned. |
   | maintenance_window | True | dict |  | Starting time of a weekly 4 hour-long window, during which maintenance might occur in hh:mm:ss format |
   | api_url | False | str |  | The Ionos API base URL. |
