+++ docs/api/dataplatform/dataplatform_cluster_config.md
@@ -1,4 +1,4 @@
-# data_platform_cluster_config
+# dataplatform_cluster_config
 
 This is a simple module that supports getting config of DataPlatform clusters
 
@@ -8,8 +8,8 @@ This is a simple module that supports getting config of DataPlatform clusters
 ```yaml
 
   - name: Get DataPlatform config
-  data_platform_cluster_config:
-    data_platform_cluster_id: "ed67d8b3-63c2-4abe-9bf0-073cee7739c9"
+  dataplatform_cluster_config:
+    dataplatform_cluster_id: "ed67d8b3-63c2-4abe-9bf0-073cee7739c9"
     config_file: 'config.yaml'
   
 ```
@@ -21,8 +21,8 @@ This is a simple module that supports getting config of DataPlatform clusters
 ```yaml
   
   - name: Get DataPlatform config
-  data_platform_cluster_config:
-    data_platform_cluster_id: "ed67d8b3-63c2-4abe-9bf0-073cee7739c9"
+  dataplatform_cluster_config:
+    dataplatform_cluster_id: "ed67d8b3-63c2-4abe-9bf0-073cee7739c9"
     config_file: 'config.yaml'
   
 ```
@@ -31,7 +31,7 @@ This is a simple module that supports getting config of DataPlatform clusters
 
   | Name | Required | Type | Default | Description |
   | :--- | :---: | :--- | :--- | :--- |
-  | data_platform_cluster_id | True | str |  | The ID of the Data Platform cluster. |
+  | dataplatform_cluster_id | True | str |  | The ID of the Data Platform cluster. |
   | config_file | True | str |  | The name of the file in which to save the config. |
   | api_url | False | str |  | The Ionos API base URL. |
   | username | False | str |  | The Ionos username. Overrides the IONOS_USERNAME environment variable. |
