+++ docs_generator.py
@@ -104,11 +104,11 @@ modules_to_generate = [
     's3key_info',
     'share',
     'user',
-    'data_platform_cluster',
-    'data_platform_cluster_config',
-    'data_platform_cluster_info',
-    'data_platform_nodepool',
-    'data_platform_nodepool_info',
+    'dataplatform_cluster',
+    'dataplatform_cluster_config',
+    'dataplatform_cluster_info',
+    'dataplatform_nodepool',
+    'dataplatform_nodepool_info',
     'certificate',
     'certificate_info'
 ]
