--- plugins/modules/data_platform_cluster_config.py
@@ -1,224 +1,0 @@
-import copy
-import yaml
-
-HAS_SDK = True
-try:
-    import ionoscloud_dataplatform
-except ImportError:
-    HAS_SDK = False
-
-
-from ansible import __version__
-from ansible.module_utils.basic import AnsibleModule, env_fallback
-from ansible.module_utils._text import to_native
-
-
-ANSIBLE_METADATA = {
-    'metadata_version': '1.1',
-    'status': ['preview'],
-    'supported_by': 'community',
-}
-DATAPLATFORM_USER_AGENT = 'ansible-module/%s_ionos-cloud-sdk-python/%s' % (
-__version__, ionoscloud_dataplatform.__version__)
-DOC_DIRECTORY = 'data-platform'
-STATES = ['present']
-OBJECT_NAME = 'DataPlatform Cluster Config'
-
-OPTIONS = {
-    'data_platform_cluster_id': {
-        'description': ['The ID of the Data Platform cluster.'],
-        'available': STATES,
-        'required': STATES,
-        'type': 'str',
-    },
-    'config_file': {
-        'description': ['The name of the file in which to save the config.'],
-        'available': STATES,
-        'required': STATES,
-        'type': 'str',
-    },
-    'api_url': {
-        'description': ['The Ionos API base URL.'],
-        'version_added': '2.4',
-        'env_fallback': 'IONOS_API_URL',
-        'available': STATES,
-        'type': 'str',
-    },
-    'username': {
-        # Required if no token, checked manually
-        'description': ['The Ionos username. Overrides the IONOS_USERNAME environment variable.'],
-        'aliases': ['subscription_user'],
-        'env_fallback': 'IONOS_USERNAME',
-        'available': STATES,
-        'type': 'str',
-    },
-    'password': {
-        # Required if no token, checked manually
-        'description': ['The Ionos password. Overrides the IONOS_PASSWORD environment variable.'],
-        'aliases': ['subscription_password'],
-        'available': STATES,
-        'no_log': True,
-        'env_fallback': 'IONOS_PASSWORD',
-        'type': 'str',
-    },
-    'token': {
-        # If provided, then username and password no longer required
-        'description': ['The Ionos token. Overrides the IONOS_TOKEN environment variable.'],
-        'available': STATES,
-        'no_log': True,
-        'env_fallback': 'IONOS_TOKEN',
-        'type': 'str',
-    },
-    'state': {
-        'description': ['Indicate desired state of the resource.'],
-        'default': 'present',
-        'choices': STATES,
-        'available': STATES,
-        'type': 'str',
-    },
-}
-
-def transform_for_documentation(val):
-    val['required'] = len(val.get('required', [])) == len(STATES)
-    del val['available']
-    del val['type']
-    return val
-
-DOCUMENTATION = '''
----
-module: data_platform_cluster_config
-short_description: Get DataPlatform Cluster configs
-description:
-     - This is a simple module that supports getting config of DataPlatform clusters
-version_added: "2.0"
-options:
-''' + '  ' + yaml.dump(yaml.safe_load(str({k: transform_for_documentation(v) for k, v in copy.deepcopy(OPTIONS).items()})), default_flow_style=False).replace('\n', '\n  ') + '''
-requirements:
-    - "python >= 2.6"
-    - "ionoscloud-dataplatform >= 1.0.0"
-author:
-    - "IONOS Cloud SDK Team <sdk-tooling@ionos.com>"
-'''
-
-EXAMPLE_PER_STATE = {
-  'present' : '''
-  - name: Get DataPlatform config
-  data_platform_cluster_config:
-    data_platform_cluster_id: "ed67d8b3-63c2-4abe-9bf0-073cee7739c9"
-    config_file: 'config.yaml'
-  ''',
-}
-
-EXAMPLES = '\n'.join(EXAMPLE_PER_STATE.values())
-
-
-def get_config(module, client):
-    data_platform_cluster_id = module.params.get('data_platform_cluster_id')
-    config_file = module.params.get('config_file')
-
-    try:
-        with open(config_file, 'w') as f:
-            response = ionoscloud_dataplatform.DataPlatformClusterApi(api_client=client).get_cluster_kubeconfig(cluster_id=data_platform_cluster_id)
-            f.write(response)
-
-    except Exception as e:
-        module.fail_json(msg="failed to get the k8s cluster config: %s" % to_native(e))
-
-    return {
-        'failed': False,
-        'changed': True,
-        'config': response
-    }
-
-
-def get_module_arguments():
-    arguments = {}
-
-    for option_name, option in OPTIONS.items():
-      arguments[option_name] = {
-        'type': option['type'],
-      }
-      for key in ['choices', 'default', 'aliases', 'no_log', 'elements']:
-        if option.get(key) is not None:
-          arguments[option_name][key] = option.get(key)
-
-      if option.get('env_fallback'):
-        arguments[option_name]['fallback'] = (env_fallback, [option['env_fallback']])
-
-      if len(option.get('required', [])) == len(STATES):
-        arguments[option_name]['required'] = True
-
-    return arguments
-
-
-def get_sdk_config(module, sdk):
-    username = module.params.get('username')
-    password = module.params.get('password')
-    token = module.params.get('token')
-    api_url = module.params.get('api_url')
-
-    if token is not None:
-        # use the token instead of username & password
-        conf = {
-            'token': token
-        }
-    else:
-        # use the username & password
-        conf = {
-            'username': username,
-            'password': password,
-        }
-
-    if api_url is not None:
-        conf['host'] = api_url
-        conf['server_index'] = None
-
-    return sdk.Configuration(**conf)
-
-
-def check_required_arguments(module, state, object_name):
-    # manually checking if token or username & password provided
-    if (
-        not module.params.get("token")
-        and not (module.params.get("username") and module.params.get("password"))
-    ):
-        module.fail_json(
-            msg='Token or username & password are required for {object_name} state {state}'.format(
-                object_name=object_name,
-                state=state,
-            ),
-        )
-
-    for option_name, option in OPTIONS.items():
-        if state in option.get('required', []) and not module.params.get(option_name):
-            module.fail_json(
-                msg='{option_name} parameter is required for {object_name} state {state}'.format(
-                    option_name=option_name,
-                    object_name=object_name,
-                    state=state,
-                ),
-            )
-
-
-def main():
-    module = AnsibleModule(argument_spec=get_module_arguments(), supports_check_mode=True)
-
-    if not HAS_SDK:
-        module.fail_json(msg='ionoscloud is required for this module, run `pip install ionoscloud`')
-
-    state = module.params.get('state')
-
-    dataplatform_api_client = ionoscloud_dataplatform.ApiClient(get_sdk_config(module, ionoscloud_dataplatform))
-    dataplatform_api_client.user_agent = DATAPLATFORM_USER_AGENT
-
-    check_required_arguments(module, state, OBJECT_NAME)
-
-    try:
-        if state == 'present':
-            module.exit_json(**get_config(module, dataplatform_api_client))
-    except Exception as e:
-        module.fail_json(msg='failed to set {object_name} state {state}: {error}'.format(object_name=OBJECT_NAME, error=to_native(e), state=state))
-
-
-if __name__ == '__main__':
-    main()
