+++ plugins/modules/dataplatform_cluster.py
@@ -9,9 +9,9 @@ import yaml
 HAS_SDK = True
 
 try:
-    import ionoscloud_dsaas
-    from ionoscloud_dsaas import __version__ as sdk_version
-    from ionoscloud_dsaas import ApiClient
+    import ionoscloud_dataplatform
+    from ionoscloud_dataplatform import __version__ as sdk_version
+    from ionoscloud_dataplatform import ApiClient
 except ImportError:
     HAS_SDK = False
 
@@ -21,8 +21,8 @@ ANSIBLE_METADATA = {
     'status': ['preview'],
     'supported_by': 'community',
 }
-DSAAS_USER_AGENT = 'ansible-module/%s_ionos-cloud-sdk-python-dsaas/%s' % ( __version__, sdk_version)
-DOC_DIRECTORY = 'data-platform'
+DATAPLATFORM_USER_AGENT = 'ansible-module/%s_ionos-cloud-sdk-python-dataplatform/%s' % ( __version__, sdk_version)
+DOC_DIRECTORY = 'dataplatform'
 STATES = ['present', 'absent', 'update']
 OBJECT_NAME = 'Data Platform Cluster'
 
@@ -37,13 +37,13 @@ OPTIONS = {
         'required': ['present', 'update'],
         'type': 'str',
     },
-    'data_platform_cluster_id': {
+    'dataplatform_cluster_id': {
         'description': ['The ID of the Data Platform cluster.'],
         'available': ['update', 'absent'],
         'required': ['update', 'absent'],
         'type': 'str',
     },
-    'data_platform_version': {
+    'dataplatform_version': {
         'description': ['The version of the DataPlatform.'],
         'available': ['present', 'update'],
         'required': ['update'],
@@ -126,7 +126,7 @@ def transform_for_documentation(val):
 
 DOCUMENTATION = '''
 ---
-module: data_platform_cluster
+module: dataplatform_cluster
 short_description: Create or destroy a Data Platform Cluster.
 description:
      - This is a simple module that supports creating or removing Data Platform Clusters.
@@ -136,7 +136,7 @@ options:
 ''' + '  ' + yaml.dump(yaml.safe_load(str({k: transform_for_documentation(v) for k, v in copy.deepcopy(OPTIONS).items()})), default_flow_style=False).replace('\n', '\n  ') + '''
 requirements:
     - "python >= 2.6"
-    - "ionoscloud_dsaas >= 1.0.0"
+    - "ionoscloud_dataplatform >= 1.0.0"
 author:
     - "IONOS Cloud SDK Team <sdk-tooling@ionos.com>"
 '''
@@ -144,23 +144,23 @@ author:
 EXAMPLE_PER_STATE = {
   'present' : '''
   - name: Create Data Platform cluster
-    data_platform_cluster:
+    dataplatform_cluster:
       name: "{{ cluster_name }}"
   ''',
   'update' : '''
   - name: Update Data Platform cluster
-    data_platform_cluster:
-      data_platform_cluster_id: "89a5aeb0-d6c1-4cef-8f6b-2b9866d85850"
+    dataplatform_cluster:
+      dataplatform_cluster_id: "89a5aeb0-d6c1-4cef-8f6b-2b9866d85850"
       maintenance_window:
         day_of_the_week: 'Tuesday'
         time: '13:03:00'
-      data_platform_version: 1.17.8
+      dataplatform_version: 1.17.8
       state: update
   ''',
   'absent' : '''
   - name: Delete Data Platform cluster
-    data_platform_cluster:
-      data_platform_cluster_id: "a9b56a4b-8033-4f1a-a59d-cfea86cfe40b"
+    dataplatform_cluster:
+      dataplatform_cluster_id: "a9b56a4b-8033-4f1a-a59d-cfea86cfe40b"
       state: absent
   ''',
 }
@@ -217,9 +217,9 @@ def _get_request_id(headers):
                         "header 'location': '{location}'".format(location=headers['location']))
 
 
-def create_data_platform_cluster(module, client):
+def create_dataplatform_cluster(module, client):
     cluster_name = module.params.get('cluster_name')
-    data_platform_version = module.params.get('data_platform_version')
+    dataplatform_version = module.params.get('dataplatform_version')
     maintenance = module.params.get('maintenance_window')
     datacenter_id = module.params.get('datacenter_id')
     wait = module.params.get('wait')
@@ -229,9 +229,9 @@ def create_data_platform_cluster(module, client):
         maintenance_window = dict(maintenance)
         maintenance_window['dayOfTheWeek'] = maintenance_window.pop('day_of_the_week')
 
-    dsaas_cluster_server = ionoscloud_dsaas.DataPlatformClusterApi(api_client=client)
+    dataplatform_cluster_server = ionoscloud_dataplatform.DataPlatformClusterApi(api_client=client)
 
-    existing_cluster = get_resource(module, dsaas_cluster_server.get_clusters(), cluster_name)
+    existing_cluster = get_resource(module, dataplatform_cluster_server.get_clusters(), cluster_name)
 
     if module.check_mode:
         module.exit_json(changed=False)
@@ -245,19 +245,19 @@ def create_data_platform_cluster(module, client):
         }
 
     try:
-        dsaas_cluster_properties = ionoscloud_dsaas.CreateClusterProperties(
+        dataplatform_cluster_properties = ionoscloud_dataplatform.CreateClusterProperties(
             name=cluster_name,
-            data_platform_version=data_platform_version,
+            data_platform_version=dataplatform_version,
             datacenter_id=datacenter_id,
             maintenance_window=maintenance_window,
         )
-        dsaas_cluster = ionoscloud_dsaas.CreateClusterRequest(properties=dsaas_cluster_properties)
+        dataplatform_cluster = ionoscloud_dataplatform.CreateClusterRequest(properties=dataplatform_cluster_properties)
 
-        dsaas_cluster_response = dsaas_cluster_server.create_cluster(create_cluster_request=dsaas_cluster)
+        dataplatform_cluster_response = dataplatform_cluster_server.create_cluster(create_cluster_request=dataplatform_cluster)
 
         if wait:
             client.wait_for(
-                fn_request=lambda: dsaas_cluster_server.get_clusters(),
+                fn_request=lambda: dataplatform_cluster_server.get_clusters(),
                 fn_check=lambda r: list(filter(
                     lambda e: e.properties.name == cluster_name,
                     r.items
@@ -269,7 +269,7 @@ def create_data_platform_cluster(module, client):
             'changed': True,
             'failed': False,
             'action': 'create',
-            'data_platform_cluster': dsaas_cluster_response.to_dict()
+            'dataplatform_cluster': dataplatform_cluster_response.to_dict()
         }
 
         return results
@@ -279,29 +279,29 @@ def create_data_platform_cluster(module, client):
             msg="failed to create the k8s cluster: %s" % to_native(e))
 
 
-def delete_data_platform_cluster(module, client):
-    cluster_id = module.params.get('data_platform_cluster_id')
+def delete_dataplatform_cluster(module, client):
+    cluster_id = module.params.get('dataplatform_cluster_id')
     cluster_name = module.params.get('cluster_name')
     wait = module.params.get('wait')
     changed = False
 
-    dsaas_cluster_server = ionoscloud_dsaas.DataPlatformClusterApi(api_client=client)
-    data_platform_clusters = dsaas_cluster_server.get_clusters()
+    dataplatform_cluster_server = ionoscloud_dataplatform.DataPlatformClusterApi(api_client=client)
+    dataplatform_clusters = dataplatform_cluster_server.get_clusters()
 
-    dsaas_cluster = get_resource(module, data_platform_clusters, cluster_id if cluster_id else cluster_name)
+    dataplatform_cluster = get_resource(module, dataplatform_clusters, cluster_id if cluster_id else cluster_name)
 
-    if not dsaas_cluster:
+    if not dataplatform_cluster:
         module.exit_json(changed=False)
 
     try:
-        if dsaas_cluster.metadata.state == 'AVAILABLE':
-            dsaas_cluster_server.delete_cluster(cluster_id=dsaas_cluster.id)
+        if dataplatform_cluster.metadata.state == 'AVAILABLE':
+            dataplatform_cluster_server.delete_cluster(cluster_id=dataplatform_cluster.id)
 
         if wait:
             client.wait_for(
-                fn_request=lambda: dsaas_cluster_server.get_clusters(),
+                fn_request=lambda: dataplatform_cluster_server.get_clusters(),
                 fn_check=lambda r: len(list(filter(
-                    lambda e: e.id == dsaas_cluster.id,
+                    lambda e: e.id == dataplatform_cluster.id,
                     r.items
                 ))) < 1,
                 console_print='.',
@@ -315,40 +315,40 @@ def delete_data_platform_cluster(module, client):
     return {
         'action': 'delete',
         'changed': changed,
-        'id': dsaas_cluster.id,
+        'id': dataplatform_cluster.id,
     }
 
 
-def update_k8s_cluster(module, client):
+def update_dataplatform_cluster(module, client):
     cluster_name = module.params.get('cluster_name')
-    data_platform_version = module.params.get('data_platform_version')
-    data_platform_cluster_id = module.params.get('data_platform_cluster_id')
+    dataplatform_version = module.params.get('dataplatform_version')
+    dataplatform_cluster_id = module.params.get('dataplatform_cluster_id')
     maintenance = module.params.get('maintenance_window')
 
     maintenance_window = dict(maintenance)
     maintenance_window['dayOfTheWeek'] = maintenance_window.pop('day_of_the_week')
 
-    dsaas_cluster_server = ionoscloud_dsaas.DataPlatformClusterApi(api_client=client)
+    dataplatform_cluster_server = ionoscloud_dataplatform.DataPlatformClusterApi(api_client=client)
     
-    existing_cluster_id_by_name = get_resource_id(module, dsaas_cluster_server.get_clusters(), cluster_name)
+    existing_cluster_id_by_name = get_resource_id(module, dataplatform_cluster_server.get_clusters(), cluster_name)
 
-    if data_platform_cluster_id is not None and existing_cluster_id_by_name is not None and existing_cluster_id_by_name != k8s_cluster_id:
+    if dataplatform_cluster_id is not None and existing_cluster_id_by_name is not None and existing_cluster_id_by_name != k8s_cluster_id:
             module.fail_json(msg='failed to update the {}: Another resource with the desired name ({}) exists'.format(OBJECT_NAME, cluster_name))
 
     if module.check_mode:
         module.exit_json(changed=True)
     try:
-        dsaas_cluster_properties = ionoscloud_dsaas.PatchClusterProperties(
+        dataplatform_cluster_properties = ionoscloud_dataplatform.PatchClusterProperties(
             name=cluster_name,
-            data_platform_version=data_platform_version,
+            data_platform_version=dataplatform_version,
             maintenance_window=maintenance_window,
         )
-        dsaas_cluster = ionoscloud_dsaas.PatchClusterRequest(properties=dsaas_cluster_properties)
-        dsaas_cluster = dsaas_cluster_server.patch_cluster(data_platform_cluster_id, dsaas_cluster)
+        dataplatform_cluster = ionoscloud_dataplatform.PatchClusterRequest(properties=dataplatform_cluster_properties)
+        dataplatform_cluster = dataplatform_cluster_server.patch_cluster(dataplatform_cluster_id, dataplatform_cluster)
 
         if module.params.get('wait'):
             client.wait_for(
-                fn_request=lambda: dsaas_cluster_server.get_clusters(),
+                fn_request=lambda: dataplatform_cluster_server.get_clusters(),
                 fn_check=lambda r: list(filter(
                     lambda e: e.properties.name == cluster_name,
                     r.items
@@ -365,7 +365,7 @@ def update_k8s_cluster(module, client):
         'changed': changed,
         'failed': False,
         'action': 'update',
-        'data_platform_cluster': dsaas_cluster.to_dict()
+        'dataplatform_cluster': dataplatform_cluster.to_dict()
     }
 
 
@@ -442,20 +442,20 @@ def main():
     module = AnsibleModule(argument_spec=get_module_arguments(), supports_check_mode=True)
 
     if not HAS_SDK:
-        module.fail_json(msg='ionoscloud_dsaas is required for this module, run `pip install ionoscloud_dsaas`')
+        module.fail_json(msg='ionoscloud_dataplatform is required for this module, run `pip install ionoscloud_dataplatform`')
 
     state = module.params.get('state')
-    with ApiClient(get_sdk_config(module, ionoscloud_dsaas)) as api_client:
-        api_client.user_agent = DSAAS_USER_AGENT
+    with ApiClient(get_sdk_config(module, ionoscloud_dataplatform)) as api_client:
+        api_client.user_agent = DATAPLATFORM_USER_AGENT
         check_required_arguments(module, state, OBJECT_NAME)
 
         try:
             if state == 'present':
-                module.exit_json(**create_data_platform_cluster(module, api_client))
+                module.exit_json(**create_dataplatform_cluster(module, api_client))
             elif state == 'absent':
-                module.exit_json(**delete_data_platform_cluster(module, api_client))
+                module.exit_json(**delete_dataplatform_cluster(module, api_client))
             elif state == 'update':
-                module.exit_json(**update_data_platform_cluster(module, api_client))
+                module.exit_json(**update_dataplatform_cluster(module, api_client))
         except Exception as e:
             module.fail_json(msg='failed to set {object_name} state {state}: {error}'.format(object_name=OBJECT_NAME, error=to_native(e), state=state))
