+++ plugins/modules/dataplatform_cluster_config.py
@@ -1,15 +1,17 @@
 import copy
 import yaml
 
-from ansible import __version__
-from ansible.module_utils.basic import AnsibleModule, env_fallback
-from ansible.module_utils._text import to_native
-
 HAS_SDK = True
 try:
     import ionoscloud_dataplatform
 except ImportError:
     HAS_SDK = False
+
+
+from ansible import __version__
+from ansible.module_utils.basic import AnsibleModule, env_fallback
+from ansible.module_utils._text import to_native
+
 
 ANSIBLE_METADATA = {
     'metadata_version': '1.1',
@@ -18,13 +20,19 @@ ANSIBLE_METADATA = {
 }
 DATAPLATFORM_USER_AGENT = 'ansible-module/%s_ionos-cloud-sdk-python/%s' % (
 __version__, ionoscloud_dataplatform.__version__)
-DOC_DIRECTORY = 'data-platform'
-STATES = ['info']
-OBJECT_NAME = 'DataPlatform Clusters'
+DOC_DIRECTORY = 'dataplatform'
+STATES = ['present']
+OBJECT_NAME = 'DataPlatform Cluster Config'
 
 OPTIONS = {
-    'data_platform_cluster_id': {
+    'dataplatform_cluster_id': {
         'description': ['The ID of the Data Platform cluster.'],
+        'available': STATES,
+        'required': STATES,
+        'type': 'str',
+    },
+    'config_file': {
+        'description': ['The name of the file in which to save the config.'],
         'available': STATES,
         'required': STATES,
         'type': 'str',
@@ -61,8 +69,14 @@ OPTIONS = {
         'env_fallback': 'IONOS_TOKEN',
         'type': 'str',
     },
+    'state': {
+        'description': ['Indicate desired state of the resource.'],
+        'default': 'present',
+        'choices': STATES,
+        'available': STATES,
+        'type': 'str',
+    },
 }
-
 
 def transform_for_documentation(val):
     val['required'] = len(val.get('required', [])) == len(STATES)
@@ -70,18 +84,15 @@ def transform_for_documentation(val):
     del val['type']
     return val
 
-
 DOCUMENTATION = '''
 ---
-module: dataplatform_nodepool_info
-short_description: List DataPlatform Nodepools
+module: dataplatform_cluster_config
+short_description: Get DataPlatform Cluster configs
 description:
-     - This is a simple module that supports listing existing DataPlatform Nodepools
+     - This is a simple module that supports getting config of DataPlatform clusters
 version_added: "2.0"
 options:
-''' + '  ' + yaml.dump(
-    yaml.safe_load(str({k: transform_for_documentation(v) for k, v in copy.deepcopy(OPTIONS).items()})),
-    default_flow_style=False).replace('\n', '\n  ') + '''
+''' + '  ' + yaml.dump(yaml.safe_load(str({k: transform_for_documentation(v) for k, v in copy.deepcopy(OPTIONS).items()})), default_flow_style=False).replace('\n', '\n  ') + '''
 requirements:
     - "python >= 2.6"
     - "ionoscloud-dataplatform >= 1.0.0"
@@ -89,34 +100,53 @@ author:
     - "IONOS Cloud SDK Team <sdk-tooling@ionos.com>"
 '''
 
-EXAMPLES = '''
-    - name: List DataPlatform Nodepools
-        data_platform_nodepool_info:
-        register: data_platform_nodepools_response
+EXAMPLE_PER_STATE = {
+  'present' : '''
+  - name: Get DataPlatform config
+  dataplatform_cluster_config:
+    dataplatform_cluster_id: "ed67d8b3-63c2-4abe-9bf0-073cee7739c9"
+    config_file: 'config.yaml'
+  ''',
+}
+
+EXAMPLES = '\n'.join(EXAMPLE_PER_STATE.values())
 
 
-    - name: Show DataPlatform Clusters
-        debug:
-            var: data_platform_nodepools_response.result
-'''
+def get_config(module, client):
+    dataplatform_cluster_id = module.params.get('dataplatform_cluster_id')
+    config_file = module.params.get('config_file')
+
+    try:
+        with open(config_file, 'w') as f:
+            response = ionoscloud_dataplatform.DataPlatformClusterApi(api_client=client).get_cluster_kubeconfig(cluster_id=dataplatform_cluster_id)
+            f.write(response)
+
+    except Exception as e:
+        module.fail_json(msg="failed to get the k8s cluster config: %s" % to_native(e))
+
+    return {
+        'failed': False,
+        'changed': True,
+        'config': response
+    }
 
 
 def get_module_arguments():
     arguments = {}
 
     for option_name, option in OPTIONS.items():
-        arguments[option_name] = {
-            'type': option['type'],
-        }
-        for key in ['choices', 'default', 'aliases', 'no_log', 'elements']:
-            if option.get(key) is not None:
-                arguments[option_name][key] = option.get(key)
+      arguments[option_name] = {
+        'type': option['type'],
+      }
+      for key in ['choices', 'default', 'aliases', 'no_log', 'elements']:
+        if option.get(key) is not None:
+          arguments[option_name][key] = option.get(key)
 
-        if option.get('env_fallback'):
-            arguments[option_name]['fallback'] = (env_fallback, [option['env_fallback']])
+      if option.get('env_fallback'):
+        arguments[option_name]['fallback'] = (env_fallback, [option['env_fallback']])
 
-        if len(option.get('required', [])) == len(STATES):
-            arguments[option_name]['required'] = True
+      if len(option.get('required', [])) == len(STATES):
+        arguments[option_name]['required'] = True
 
     return arguments
 
@@ -146,24 +176,26 @@ def get_sdk_config(module, sdk):
     return sdk.Configuration(**conf)
 
 
-def check_required_arguments(module, object_name):
+def check_required_arguments(module, state, object_name):
     # manually checking if token or username & password provided
     if (
-            not module.params.get("token")
-            and not (module.params.get("username") and module.params.get("password"))
+        not module.params.get("token")
+        and not (module.params.get("username") and module.params.get("password"))
     ):
         module.fail_json(
-            msg='Token or username & password are required for {object_name}'.format(
+            msg='Token or username & password are required for {object_name} state {state}'.format(
                 object_name=object_name,
+                state=state,
             ),
         )
 
     for option_name, option in OPTIONS.items():
-        if 'info' in option.get('required', []) and not module.params.get(option_name):
+        if state in option.get('required', []) and not module.params.get(option_name):
             module.fail_json(
-                msg='{option_name} parameter is required for retrieving {object_name}'.format(
+                msg='{option_name} parameter is required for {object_name} state {state}'.format(
                     option_name=option_name,
                     object_name=object_name,
+                    state=state,
                 ),
             )
 
@@ -172,22 +204,20 @@ def main():
     module = AnsibleModule(argument_spec=get_module_arguments(), supports_check_mode=True)
 
     if not HAS_SDK:
-        module.fail_json(
-            msg='ionoscloud_dataplatform is required for this module, run `pip install ionoscloud_dataplatform`')
+        module.fail_json(msg='ionoscloud is required for this module, run `pip install ionoscloud`')
+
+    state = module.params.get('state')
 
     dataplatform_api_client = ionoscloud_dataplatform.ApiClient(get_sdk_config(module, ionoscloud_dataplatform))
     dataplatform_api_client.user_agent = DATAPLATFORM_USER_AGENT
 
-    check_required_arguments(module, OBJECT_NAME)
+    check_required_arguments(module, state, OBJECT_NAME)
+
     try:
-        data_platform_cluster_id = module.params.get('data_platform_cluster_id')
-        results = []
-        for nodepool in ionoscloud_dataplatform.DataPlatformNodePoolApi(dataplatform_api_client).get_cluster_nodepools(cluster_id=data_platform_cluster_id).items:
-            results.append(nodepool.to_dict())
-        module.exit_json(result=results)
+        if state == 'present':
+            module.exit_json(**get_config(module, dataplatform_api_client))
     except Exception as e:
-        module.fail_json(
-            msg='failed to retrieve {object_name}: {error}'.format(object_name=OBJECT_NAME, error=to_native(e)))
+        module.fail_json(msg='failed to set {object_name} state {state}: {error}'.format(object_name=OBJECT_NAME, error=to_native(e), state=state))
 
 
 if __name__ == '__main__':
