+++ plugins/modules/dataplatform_nodepool.py
@@ -7,9 +7,9 @@ import yaml
 HAS_SDK = True
 
 try:
-    import ionoscloud_dsaas
-    from ionoscloud_dsaas import __version__ as sdk_version
-    from ionoscloud_dsaas import ApiClient
+    import ionoscloud_dataplatform
+    from ionoscloud_dataplatform import __version__ as sdk_version
+    from ionoscloud_dataplatform import ApiClient
 except ImportError:
     HAS_SDK = False
 
@@ -23,8 +23,8 @@ ANSIBLE_METADATA = {
     'status': ['preview'],
     'supported_by': 'community',
 }
-DSAAS_USER_AGENT = 'ansible-module/%s_ionos-cloud-sdk-python-dsaas/%s' % ( __version__, sdk_version)
-DOC_DIRECTORY = 'data-platform'
+DATAPLATFORM_USER_AGENT = 'ansible-module/%s_ionos-cloud-sdk-python-dataplatform/%s' % ( __version__, sdk_version)
+DOC_DIRECTORY = 'dataplatform'
 STATES = ['present', 'absent', 'update']
 OBJECT_NAME = 'Data Platform Nodepool'
 
@@ -39,13 +39,13 @@ OPTIONS = {
         'required': ['present'],
         'type': 'str',
     },
-    'data_platform_cluster_id': {
+    'dataplatform_cluster_id': {
         'description': ['The ID of the Data Platform cluster.'],
         'available': STATES,
         'required': STATES,
         'type': 'str',
     },
-    'data_platform_nodepool_id': {
+    'dataplatform_nodepool_id': {
         'description': ['The ID of the Data Platform nodepool.'],
         'available': ['update', 'absent'],
         'required': ['update', 'absent'],
@@ -185,27 +185,27 @@ def transform_for_documentation(val):
 
 DOCUMENTATION = '''
 ---
-module: data_platform_nodepool
+module: dataplatform_nodepool
 short_description: Create or destroy Data Platform Nodepools
 description:
      - This is a simple module that supports creating or removing Data Platform Nodepools.
-       This module has a dependency on ionoscloud_dsaas >= 1.0.0
+       This module has a dependency on ionoscloud_dataplatform >= 1.0.0
 version_added: "2.0"
 options:
 ''' + '  ' + yaml.dump(yaml.safe_load(str({k: transform_for_documentation(v) for k, v in copy.deepcopy(OPTIONS).items()})), default_flow_style=False).replace('\n', '\n  ') + '''
 requirements:
     - "python >= 2.6"
-    - "ionoscloud_dsaas >= 1.0.0"
+    - "ionoscloud_dataplatform >= 1.0.0"
 author:
     - "IONOS Cloud SDK Team <sdk-tooling@ionos.com>"
 '''
 
 EXAMPLE_PER_STATE = {
   'present' : '''
-  - name: Create Dsaas Platform nodepool
-    data_platform_nodepool:
+  - name: Create Data Platform nodepool
+    dataplatform_nodepool:
       cluster_name: "{{ name }}"
-      data_platform_cluster_id: "a0a65f51-4d3c-438c-9543-39a3d7668af3"
+      dataplatform_cluster_id: "a0a65f51-4d3c-438c-9543-39a3d7668af3"
       node_count: "1"
       cpu_family: "AMD_OPTERON"
       cores_count: "1"
@@ -215,10 +215,10 @@ EXAMPLE_PER_STATE = {
       storage_size: "100"
   ''',
   'update' : '''
-  - name: Update Dsaas Platform nodepool
-    data_platform_nodepool:
+  - name: Update Data Platform nodepool
+    dataplatform_nodepool:
       cluster_name: "{{ name }}"
-      data_platform_cluster_id: "ed67d8b3-63c2-4abe-9bf0-073cee7739c9"
+      dataplatform_cluster_id: "ed67d8b3-63c2-4abe-9bf0-073cee7739c9"
       node_count: 1
       cores_count: "1"
       maintenance_window:
@@ -230,10 +230,10 @@ EXAMPLE_PER_STATE = {
       state: update
   ''',
   'absent' : '''
-  - name: Delete Dsaas Platform nodepool
-    data_platform_nodepool:
-      data_platform_cluster_id: "a0a65f51-4d3c-438c-9543-39a3d7668af3"
-      data_platform_nodepool_id: "e3aa6101-436f-49fa-9a8c-0d6617e0a277"
+  - name: Delete Data Platform nodepool
+    dataplatform_nodepool:
+      dataplatform_cluster_id: "a0a65f51-4d3c-438c-9543-39a3d7668af3"
+      dataplatform_nodepool_id: "e3aa6101-436f-49fa-9a8c-0d6617e0a277"
       state: absent
   ''',
 }
@@ -280,9 +280,9 @@ def get_resource_id(module, resource_list, identity, identity_paths=None):
     return resource.id if resource is not None else None
 
 
-def create_data_platform_nodepool(module, client):
-    data_platform_cluster_id = module.params.get('data_platform_cluster_id')
-    data_platform_version = module.params.get('data_platform_version')
+def create_dataplatform_nodepool(module, client):
+    dataplatform_cluster_id = module.params.get('dataplatform_cluster_id')
+    dataplatform_version = module.params.get('dataplatform_version')
     nodepool_name = module.params.get('nodepool_name')
     node_count = module.params.get('node_count')
     cpu_family = module.params.get('cpu_family')
@@ -296,14 +296,14 @@ def create_data_platform_nodepool(module, client):
     annotations = module.params.get('annotations')
     wait = module.params.get('wait')
 
-    dsaas_nodepool_server = ionoscloud_dsaas.DataPlatformNodePoolApi(api_client=client)
+    dataplatform_nodepool_server = ionoscloud_dataplatform.DataPlatformNodePoolApi(api_client=client)
 
     maintenance_window = None
     if maintenance:
         maintenance_window = dict(maintenance)
         maintenance_window['dayOfTheWeek'] = maintenance_window.pop('day_of_the_week')
 
-    nodepool_list = dsaas_nodepool_server.get_cluster_nodepools(data_platform_cluster_id)
+    nodepool_list = dataplatform_nodepool_server.get_cluster_nodepools(dataplatform_cluster_id)
 
     existing_nodepool = get_resource(module, nodepool_list, nodepool_name)
 
@@ -312,11 +312,11 @@ def create_data_platform_nodepool(module, client):
             'changed': False,
             'failed': False,
             'action': 'create',
-            'data_platform_nodepool': existing_nodepool.to_dict()
+            'dataplatform_nodepool': existing_nodepool.to_dict()
         }
 
     try:
-        data_platform_nodepool_properties = ionoscloud_dsaas.CreateNodePoolProperties(
+        dataplatform_nodepool_properties = ionoscloud_dataplatform.CreateNodePoolProperties(
             name=nodepool_name,
             node_count=node_count,
             cpu_family=cpu_family,
@@ -325,19 +325,19 @@ def create_data_platform_nodepool(module, client):
             availability_zone=availability_zone,
             storage_type=storage_type,
             storage_size=storage_size,
-            data_platform_version=data_platform_version,
+            dataplatform_version=dataplatform_version,
             maintenance_window=maintenance_window,
             labels=labels,
             annotations=annotations,
         )
 
-        dsaas_nodepool = ionoscloud_dsaas.CreateNodePoolRequest(properties=data_platform_nodepool_properties)
+        dataplatform_nodepool = ionoscloud_dataplatform.CreateNodePoolRequest(properties=dataplatform_nodepool_properties)
 
-        dsaas_response= dsaas_nodepool_server.create_cluster_nodepool(data_platform_cluster_id, dsaas_nodepool)
+        dataplatform_response= dataplatform_nodepool_server.create_cluster_nodepool(dataplatform_cluster_id, dataplatform_nodepool)
 
         if wait:
             client.wait_for(
-                fn_request=lambda: dsaas_nodepool_server.get_cluster_nodepools(data_platform_cluster_id),
+                fn_request=lambda: dataplatform_nodepool_server.get_cluster_nodepools(dataplatform_cluster_id),
                 fn_check=lambda r: list(filter(
                     lambda e: e.properties.name == nodepool_name,
                     r.items
@@ -349,7 +349,7 @@ def create_data_platform_nodepool(module, client):
             'changed': True,
             'failed': False,
             'action': 'create',
-            'data_platform_nodepool': dsaas_response.to_dict()
+            'dataplatform_nodepool': dataplatform_response.to_dict()
         }
         return results
 
@@ -357,26 +357,26 @@ def create_data_platform_nodepool(module, client):
         module.fail_json(msg="failed to create the Data Platform nodepool: %s" % to_native(e))
 
 
-def delete_data_platform_nodepool(module, client):
-    data_platform_cluster_id = module.params.get('data_platform_cluster_id')
-    nodepool_id = module.params.get('data_platform_nodepool_id')
+def delete_dataplatform_nodepool(module, client):
+    dataplatform_cluster_id = module.params.get('dataplatform_cluster_id')
+    nodepool_id = module.params.get('dataplatform_nodepool_id')
     nodepool_name = module.params.get('nodepool_name')
 
-    dsaas_nodepool_server = ionoscloud_dsaas.DataPlatformNodePoolApi(api_client=client)
-    nodepool_list = dsaas_nodepool_server.get_cluster_nodepools(data_platform_cluster_id)
-    dsaas_nodepool = get_resource(module, nodepool_list, nodepool_id if nodepool_id else nodepool_name)
+    dataplatform_nodepool_server = ionoscloud_dataplatform.DataPlatformNodePoolApi(api_client=client)
+    nodepool_list = dataplatform_nodepool_server.get_cluster_nodepools(dataplatform_cluster_id)
+    dataplatform_nodepool = get_resource(module, nodepool_list, nodepool_id if nodepool_id else nodepool_name)
 
-    if not dsaas_nodepool:
+    if not dataplatform_nodepool:
         module.exit_json(changed=False)
 
     changed = False
 
     try:
-        if dsaas_nodepool.metadata.state == 'AVAILABLE':
-            dsaas_nodepool_server.delete_cluster_nodepool(data_platform_cluster_id, nodepool_id)
+        if dataplatform_nodepool.metadata.state == 'AVAILABLE':
+            dataplatform_nodepool_server.delete_cluster_nodepool(dataplatform_cluster_id, nodepool_id)
         if module.params.get('wait'):
             client.wait_for(
-                fn_request=lambda: dsaas_nodepool_server.get_cluster_nodepools(data_platform_cluster_id),
+                fn_request=lambda: dataplatform_nodepool_server.get_cluster_nodepools(dataplatform_cluster_id),
                 fn_check=lambda r: len(list(filter(
                     lambda e: e.id == nodepool_id,
                     r.items
@@ -392,29 +392,29 @@ def delete_data_platform_nodepool(module, client):
     return {
         'action': 'delete',
         'changed': changed,
-        'data_platform_nodepool_id': nodepool_id
+        'dataplatform_nodepool_id': nodepool_id
     }
 
 
-def update_data_platform_nodepool(module, client):
-    data_platform_cluster_id = module.params.get('data_platform_cluster_id')
+def update_dataplatform_nodepool(module, client):
+    dataplatform_cluster_id = module.params.get('dataplatform_cluster_id')
     nodepool_id = module.params.get('nodepool_id')
     node_count = module.params.get('node_count')
     maintenance = module.params.get('maintenance_window')
     wait = module.params.get('wait')
     nodepool_name = module.params.get('nodepool_name')
-    data_platform_version = module.params.get('data_platform_version')
+    dataplatform_version = module.params.get('dataplatform_version')
     labels = module.params.get('labels')
     annotations = module.params.get('annotations')
 
-    dsaas_nodepool_server = ionoscloud_dsaas.DataPlatformNodePoolApi(api_client=client)
+    dataplatform_nodepool_server = ionoscloud_dataplatform.DataPlatformNodePoolApi(api_client=client)
 
     maintenance_window = None
     if maintenance:
         maintenance_window = dict(maintenance)
         maintenance_window['dayOfTheWeek'] = maintenance_window.pop('day_of_the_week')
 
-    nodepool_list = dsaas_nodepool_server.get_cluster_nodepools(data_platform_cluster_id)
+    nodepool_list = dataplatform_nodepool_server.get_cluster_nodepools(dataplatform_cluster_id)
     existing_nodepool_by_name = get_resource(module, nodepool_list, nodepool_name)
 
     if nodepool_id is not None and existing_nodepool_by_name is not None and existing_nodepool_by_name.id != nodepool_id:
@@ -428,23 +428,23 @@ def update_data_platform_nodepool(module, client):
     if module.check_mode:
         module.exit_json(changed=True)
     try:
-        data_platform_nodepool_properties = ionoscloud_dsaas.PatchNodePoolProperties(
+        dataplatform_nodepool_properties = ionoscloud_dataplatform.PatchNodePoolProperties(
             name=nodepool_name,
             node_count=node_count,
-            data_platform_version=data_platform_version,
+            dataplatform_version=dataplatform_version,
             maintenance_window=maintenance_window,
             labels=labels,
             annotations=annotations,
         )
 
-        k8s_nodepool = ionoscloud_dsaas.PatchNodePoolRequest(properties=data_platform_nodepool_properties)
-        dsaas_response = dsaas_nodepool_server.patch_cluster_nodepool(
-            data_platform_cluster_id, nodepool_id, kubernetes_node_pool=k8s_nodepool,
+        k8s_nodepool = ionoscloud_dataplatform.PatchNodePoolRequest(properties=dataplatform_nodepool_properties)
+        dataplatform_response = dataplatform_nodepool_server.patch_cluster_nodepool(
+            dataplatform_cluster_id, nodepool_id, kubernetes_node_pool=k8s_nodepool,
         )
 
         if wait:
             client.wait_for(
-                fn_request=lambda: dsaas_nodepool_server.get_cluster_nodepools(data_platform_cluster_id),
+                fn_request=lambda: dataplatform_nodepool_server.get_cluster_nodepools(dataplatform_cluster_id),
                 fn_check=lambda r: list(filter(
                     lambda e: e.id == nodepool_id,
                     r.items
@@ -462,7 +462,7 @@ def update_data_platform_nodepool(module, client):
         'changed': changed,
         'failed': False,
         'action': 'update',
-        'data_plaform_nodepool': dsaas_response.to_dict()
+        'data_plaform_nodepool': dataplatform_response.to_dict()
     }
 
 
@@ -539,11 +539,11 @@ def main():
     module = AnsibleModule(argument_spec=get_module_arguments(), supports_check_mode=True)
 
     if not HAS_SDK:
-        module.fail_json(msg='ionoscloud_dsaas is required for this module, run `pip install ionoscloud_dsaas`')
+        module.fail_json(msg='ionoscloud_dataplatform is required for this module, run `pip install ionoscloud_dataplatform`')
 
     state = module.params.get('state')
-    with ApiClient(get_sdk_config(module, ionoscloud_dsaas)) as api_client:
-        api_client.user_agent = DSAAS_USER_AGENT
+    with ApiClient(get_sdk_config(module, ionoscloud_dataplatform)) as api_client:
+        api_client.user_agent = DATAPLATFORM_USER_AGENT
         check_required_arguments(module, state, OBJECT_NAME)
         if state == 'present' and not module.params.get('node_count') and not module.params.get('auto_scaling'):
             module.fail_json(
@@ -551,11 +551,11 @@ def main():
             )
         try:
             if state == 'present':
-                module.exit_json(**create_data_platform_nodepool(module, api_client))
+                module.exit_json(**create_dataplatform_nodepool(module, api_client))
             elif state == 'absent':
-                module.exit_json(**delete_data_platform_nodepool(module, api_client))
+                module.exit_json(**delete_dataplatform_nodepool(module, api_client))
             elif state == 'update':
-                module.exit_json(**update_data_platform_nodepool(module, api_client))
+                module.exit_json(**update_dataplatform_nodepool(module, api_client))
         except Exception as e:
             module.fail_json(msg='failed to set {object_name} state {state}: {error}'.format(object_name=OBJECT_NAME, error=to_native(e), state=state))
